requirements.txt
Pillow==10.0.0
numpy==1.25.2
pytest==7.4.0